% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factbase.R
\name{AddFactbaseProvenance}
\alias{AddFactbaseProvenance}
\title{WARNING: THIS FEATURE IS INCOMPLETE.  DO NOT USE THIS FUNCTION.}
\usage{
AddFactbaseProvenance(
  x,
  description,
  hyperlink = NA_character_,
  owner = NA_character_
)
}
\arguments{
\item{x}{An object containing data that will probably eventually find its way into Factbase.}

\item{description}{A description of where `x` came from.}

\item{hyperlink}{(optional) A link to a web page where more can be read about the data.
Preferably this is a link into the system that calls this function.}

\item{owner}{The name (usually an email address) of whoever should be contacted to deal with problems
or questions about this data.}
}
\value{
`x` with provenance information added as an attribute.  The new information will be appended
  to any existing provenance.
}
\description{
Adds provenance information to data that will be uploaded to Factbase.  This should be called
in steps prior to the operation that sends the data to Factbase so that chains of calculations
can be shown to the user, and so that Factbase can determine the step at which a chain of
calculations has broken.
}
