% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\docType{data}
\name{api.root}
\alias{api.root}
\title{Constant endpoint address for accessing the Data Mart.}
\format{An object of class \code{character} of length 1.}
\usage{
api.root
}
\description{
Constant endpoint address for accessing the Data Mart.
}
\keyword{datasets}
