% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dropbox.R
\name{CheckDropboxToken}
\alias{CheckDropboxToken}
\title{Check that token is valid}
\usage{
CheckDropboxToken(token)
}
\arguments{
\item{token}{Dropbox API token to check.}
}
\value{
Returns a true if token is valid.
}
\description{
Check that token is valid
}
