% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{close.qpostcon}
\alias{close.qpostcon}
\title{Closes a QFileOpen connection}
\usage{
\method{close}{qpostcon}(con, ...)
}
\arguments{
\item{con}{connection object of class 'qpostconn'. Connection opened with QFileOpen}

\item{...}{arguments passed to or from other methods.}
}
\value{
NULL invisibly. Called for the purpose of uploading data
and assumed to succeed if no errors are thrown.
}
\description{
This is an overload for close.connection which writes the file contents 
of a connection opened using QFileOpen to the Data Mart.
}
