% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{close.qpostconn}
\alias{close.qpostconn}
\title{Closes a QFileOpen connection}
\usage{
\method{close}{qpostconn}(conn)
}
\arguments{
\item{conn}{connection object of class 'qpostconn'. Connection opened with QFileOpen}
}
\description{
This is an overload for close.connection which writes the file contents 
of a connection opened using QFileOpen to the Data Mart.
}
