% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactbaseWidget.R
\name{FactbaseUploadWidget}
\alias{FactbaseUploadWidget}
\title{This function is for Displayr internal use only.}
\usage{
FactbaseUploadWidget(
  factbase.token = "",
  mode = "Replace all",
  aggregation = "None",
  time_aggregation = NULL,
  period_type = NULL,
  definition = NULL,
  hyperlink = NULL,
  owner = NULL,
  do.upload = TRUE,
  selection.type = "Table",
  input.table = NULL,
  make.dummy.metric = NULL,
  metric.variables = NULL,
  metric.column.names = "",
  date.column.name = "",
  date.variable = NULL,
  dimension.column.names = "",
  dimension.variables = NULL,
  output.type = "Data Summaries",
  start.date = "",
  time.zone = "Australia/Sydney",
  update.period = "days",
  update.frequency = 1,
  us.format = FALSE
)
}
\arguments{
\item{factbase.token}{Undocumented}

\item{mode}{Undocumented}

\item{aggregation}{Undocumented}

\item{time_aggregation}{Undocumented}

\item{period_type}{Undocumented}

\item{definition}{Undocumented}

\item{hyperlink}{Undocumented}

\item{owner}{Undocumented}

\item{do.upload}{Undocumented}

\item{selection.type}{Undocumented}

\item{input.table}{Undocumented}

\item{make.dummy.metric}{Undocumented}

\item{metric.variables}{Undocumented}

\item{metric.column.names}{Undocumented}

\item{date.column.name}{Undocumented}

\item{date.variable}{Undocumented}

\item{dimension.column.names}{Undocumented}

\item{dimension.variables}{Undocumented}

\item{output.type}{Undocumented}

\item{start.date}{Undocumented}

\item{time.zone}{Undocumented}

\item{update.period}{Undocumented}

\item{update.frequency}{Undocumented}

\item{us.format}{Undocumented}
}
\value{
An HTMLwidget that shows summary details about the upload.
}
\description{
The R code for the standard R item *Upload to Factbase*, which provides a
zero code way to upload a metric using the Factbase R API.
}
\details{
Parameters come out of the R UI JavaScript code in
https://github.com/Displayr/factbase/blob/master/displayr/Upload%20to%20Factbase.RScript
See that code to understand the meaning of those controls.

We would prefer not to export this function, but not doing so causes build
failures.  We cannot move the function into flipFormat because it calls
functions in flipAPI that would trigger circular import dependencies because
flipAPI depends on flipFormat.
}
