% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{getApiRoot}
\alias{getApiRoot}
\title{Gets region from the environment and builds the api root. Throws an error if not found.}
\usage{
getApiRoot()
}
\value{
Region-specific api root as a string.
}
\description{
Gets region from the environment and builds the api root. Throws an error if not found.
}
