% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel.R
\name{GetDirectLink}
\alias{GetDirectLink}
\title{Converts URL from cloud storage to direct link}
\usage{
GetDirectLink(url)
}
\arguments{
\item{url}{String; url to convert}
}
\description{
Currently only supports links from Dropbox, OneDrive and Google Drive
}
