% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{getFileType}
\alias{getFileType}
\title{Guesses the type of file from the filename. Used for QSaveData/QLoadData}
\usage{
getFileType(filename)
}
\arguments{
\item{filename}{string. Filename which we are checking}
}
\value{
string or NULL. The supported file type which we have determined the file to be.
}
\description{
Guesses the type of file from the filename. Used for QSaveData/QLoadData
}
