% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{getResponseFileType}
\alias{getResponseFileType}
\title{Gets the file type of a response based on Content-Type. Used for QSaveData/QLoadData}
\usage{
getResponseFileType(response)
}
\arguments{
\item{response}{Response object.}
}
\value{
string (or NULL). The supported file type which we have determined the file to be else NULL.
}
\description{
Gets the file type of a response based on Content-Type. Used for QSaveData/QLoadData
}
