% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoogleAnalytics.R
\name{GoogleAnalytics}
\alias{GoogleAnalytics}
\title{Request data from Google Analytics}
\usage{
GoogleAnalytics(dimensions, metrics, start.date, end.date, secret.text,
  table.num = 1)
}
\arguments{
\item{dimensions}{A string (comma separated) or list of strings containing
the dimensions to request.}

\item{metrics}{Similar to \code{dimensions} but for metrics.}

\item{start.date, end.date}{The start and end dates for the query. If it is a
string it will be coerced to a date using YMD format.}

\item{secret.text}{The JSON formatted string containing the secret text that
identifies your account.}

\item{table.num}{If your account contains multiple tables, use this to select
the table to query. Defaults to the first table.}
}
\description{
Request data from Google Analytics
}
\details{
See
  \url{https://developers.google.com/analytics/devguides/reporting/core/dimsmets}
  for details of the available dimensions and metrics.

  See
  \url{http://wiki.q-researchsoftware.com/wiki/How_to_Import_Data_from_Google_Analytics}
  for more information about setting up your Google Analytics account for
  use with this function. The parameter \code{secret.text} is the text
  contained in the JSON file you get when adding a Service Account.
}

