% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dropbox.R
\name{ImportFromDropbox}
\alias{ImportFromDropbox}
\title{Import R object from file in Dropbox}
\usage{
ImportFromDropbox(importfile, token)
}
\arguments{
\item{importfile}{Name of file that the object was saved to. This is generally of the form \code{<object>.rds}.}

\item{token}{Dropbox access token with read and write access.}
}
\description{
Import R object from file in Dropbox
}
