% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{IsDisplayrCloudDriveAvailable}
\alias{IsDisplayrCloudDriveAvailable}
\title{Check that Displayr cloud drive is available by looking for the
companySecret variable.}
\usage{
IsDisplayrCloudDriveAvailable()
}
\value{
TRUE if cloud drive is available, otherwise FALSE
}
\description{
Check that Displayr cloud drive is available by looking for the
companySecret variable.
}
