% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{QFileExists}
\alias{QFileExists}
\title{Check if a file exists}
\usage{
QFileExists(filename)
}
\arguments{
\item{filename}{character string. Name of the file to search for}
}
\value{
TRUE if the file exists, otherwise FALSE
}
\description{
Check whether a file of a given name exists in the Data Mart.
}
