% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{QFileOpen}
\alias{QFileOpen}
\title{Opens a Connection}
\usage{
QFileOpen(
  filename,
  open = "r",
  blocking = TRUE,
  encoding = getOption("encoding"),
  raw = FALSE,
  method = getOption("url.method", "default"),
  mime.type = NA,
  company.token = NA
)
}
\arguments{
\item{filename}{character string. Name of file to be opened.
To reference a file in a subdirectory, use double backslashes after each folder (e.g "subdir\\file.csv").}

\item{open}{character string. See documentation for connections.}

\item{blocking}{logical. See documentation for connections.}

\item{encoding}{character string. See documentation for connections.}

\item{raw}{logical. See documentation for connections.}

\item{method}{character string. See documentation for connections.}

\item{mime.type}{character string. The mime-type of this file. If not provided, it will be interpreted from the file extension.}

\item{company.token}{Use this if you need to read from a different company's Displayr Cloud Drive. You need to contact Support to get this token.}
}
\value{
A curl connection (read) or a file connection (write)
}
\description{
Opens a connection for either reading OR writing to a file of a given name.
In the reading case, it opens a stream to the file in the Displayr Cloud Drive.
In the writing case, it opens a temporary file for writing to and uploads this to the Displayr Cloud Drive on close.
Note that writing to a file which already exists will overwrite that file's contents.
For more documentation on this function's parameters, see R documentation for opening connections.
}
