% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{QGetSharedUrl}
\alias{QGetSharedUrl}
\title{Share an object}
\usage{
QGetSharedUrl(filename)
}
\arguments{
\item{filename}{character string. Name of the file that is being shared.
To reference a file in a subdirectory, use double backslashes after each folder (e.g "subdir\\file.csv").}
}
\value{
Share URL as a string
}
\description{
Gets the cloud drive share url for an object saved to the Displayr Cloud Drive.
}
\details{
If the file has not previously been shared it will be shared else the already existing
share url will be returned.
}
