% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{QLoadData}
\alias{QLoadData}
\title{Loads an object}
\usage{
QLoadData(filename)
}
\arguments{
\item{filename}{character string. Name of the file to be opened from the Data Mart}
}
\value{
An R object
}
\description{
Loads an *.rds file from the data mart and converts this to an R object.
}
