% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{QSaveData}
\alias{QSaveData}
\title{Save an object}
\usage{
QSaveData(object, filename)
}
\arguments{
\item{object}{object. The object to be uploaded}

\item{filename}{character string. Name of the file to be written to}
}
\value{

}
\description{
Saves an object to the Data Mart as an *.rds file of the given name.
If the name is given an extension, this must be '.rds'. If there is no
extension specified, this defaults to '.rds'.
}
