% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL.R
\name{RunSQL}
\alias{RunSQL}
\title{Gets data from SQL database}
\usage{
RunSQL(
  query,
  data_provider,
  server,
  database,
  user,
  password = "",
  max_records = 100,
  port = 1433,
  warehouse = ""
)
}
\arguments{
\item{query}{SQL query to generate data returned}

\item{data_provider}{Type of database. One of "MySQL", "Postgres", "Redshift",
"Oracle", "Snowflake", "SQL Server".}

\item{server}{Host name or IP where data base is hosted}

\item{database}{Name of database}

\item{user}{User name of data base account}

\item{password}{Password of data base account}

\item{max_records}{Maximum number of rows returned in `query`}

\item{port}{Port number to connect to data base}

\item{warehouse}{Data base warehouse (only for snowflake database)}
}
\description{
Connects to SQL database and returns the result of a query
}
