% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{stopBadRequest}
\alias{stopBadRequest}
\title{Throws an error given a response. Appends error received from API if we have one.}
\usage{
stopBadRequest(obj, message = "")
}
\arguments{
\item{obj}{object. Either a bad response or a try-error.}

\item{message}{error string. The error message which will be thrown to the user.}
}
\description{
Throws an error given a response. Appends error received from API if we have one.
}
