% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{stopNotDisplayr}
\alias{stopNotDisplayr}
\title{Error when someone tries to use this package outside of Displayr.}
\usage{
stopNotDisplayr()
}
\value{
Throws an error.
}
\description{
Error when someone tries to use this package outside of Displayr.
}
