% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HubspotAPI.R
\name{timestampToDate}
\alias{timestampToDate}
\title{Convert a Hubspot timestamp to POSIXct}
\usage{
timestampToDate(timestamp)
}
\arguments{
\item{timestamp}{A timestamp in milliseconds.}
}
\description{
Convert a Hubspot timestamp (milliseconds since UNIX epoch) to POSIXct.
}

