% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateDisplayrDoc.R
\name{TriggerObjectUpdate}
\alias{TriggerObjectUpdate}
\alias{UpdateObject}
\title{Triggers a call to update R object in a Displayr document using an API post request}
\usage{
TriggerObjectUpdate(object, document.key, update.seconds = 604800)

UpdateObject(object, document.key, update.seconds = 604800)
}
\arguments{
\item{object}{String containing name of R object to export}

\item{document.key}{API key of the source document in Displayr. This can be obtained by navigating to your documents in \code{https://app.displayr.com}, and clicking on the API button next to the source document.}

\item{update.seconds}{Time in seconds after which the object in the source document will be re-updated. Note that the actual update occurs the first time the current project is open \emph{after} \code{update.seconds} has elapsed. This may occur after ImportObject is called, so that the imported object can be up to \code{\link[=ExportToDropbox]{reexport.seconds} + update.seconds} old. The default value is 1 week.}
}
\description{
Triggers a call to update R object in a Displayr document using an API post request
}
