% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factbase.R
\name{UpdateFactbasePenetrationFormula}
\alias{UpdateFactbasePenetrationFormula}
\title{Creates or updates a metric described by a formula over other metrics.
See https://factbase.azurewebsites.net/static/pages/help.html#penetration}
\usage{
UpdateFactbasePenetrationFormula(
  metric_name,
  token,
  numerator,
  denominator,
  dimensions_to_count,
  definition,
  hyperlink = NULL,
  test_return_json = F
)
}
\arguments{
\item{metric_name}{The name that will appear for selection by Factbase users.}

\item{token}{A guid that identifies and authenticates the request.  Talk to Oliver if you need
one of these.}

\item{numerator}{The name of an existing metric.  See the documentation reference above.}

\item{denominator}{The name of an existing metric.  See the documentation reference above.}

\item{dimensions_to_count}{A character vector of label dimension names.  See the documentation
reference above.}

\item{definition}{A detailed explanation of the meaning and derivation of the metric.}

\item{(Optional)}{test_return_json For testing only.  Ignore.}
}
\value{
The value of `data` that was passed in, so caller can see data uploaded if this is the
  last call in R code.
}
\description{
Creates or updates a metric described by a formula over other metrics.
See https://factbase.azurewebsites.net/static/pages/help.html#penetration
}
