% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factbase.R
\name{UpdateFactbaseRatioFormula}
\alias{UpdateFactbaseRatioFormula}
\title{Creates or updates a metric described by a formula over other metrics.
See https://factbase.azurewebsites.net/static/pages/help.html#ratio}
\usage{
UpdateFactbaseRatioFormula(
  metric_name,
  token,
  numerator,
  denominator,
  smoothing.window = NULL,
  smoothing.sum = F,
  definition,
  hyperlink = NULL,
  test_return_json = F
)
}
\arguments{
\item{metric_name}{The name that will appear for selection by Factbase users.}

\item{token}{A guid that identifies and authenticates the request.  Talk to Oliver if you need
one of these.}

\item{numerator}{The name of an existing metric.  See the documentation reference above.}

\item{denominator}{The name of an existing metric.  See the documentation reference above.}

\item{smoothing.window}{(Optional) The period over which to smooth the data.  One of "day", "week",
"month", "quarter" or "year".  See the documentation reference above.}

\item{smoothing.sum}{(Optional) TRUE to smooth using a rolling sum.  If not specified then a rolling
average will be used.  See the documentation reference above.}

\item{definition}{(Optional) A detailed explanation of the meaning and derivation of the metric.}

\item{hyperlink}{(Optional) A link to a web page where more can be read about the metric.}

\item{test_return_json}{(Optional) For testing only.  Ignore.}
}
\value{
The value of `data` that was passed in, so caller can see data uploaded if this is the
  last call in R code.
}
\description{
Creates or updates a metric described by a formula over other metrics.
See https://factbase.azurewebsites.net/static/pages/help.html#ratio
}
