% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateDashboard.R
\name{UpdateObject}
\alias{UpdateObject}
\title{Update R object in a dashboard using an API post request}
\usage{
UpdateObject(object, project.key)
}
\arguments{
\item{object}{Name of R object to export}

\item{project.key}{Secret project key of the dashboard. This can be obtained by navigating to your documents in \code{https://app.displayr.com}, and clicking on the API button next to the specified dashboard.}
}
\description{
Update R object in a dashboard using an API post request
}
