% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factbase.R
\name{UploadMetricToFactbase}
\alias{UploadMetricToFactbase}
\title{Upload a metric to Factbase.}
\usage{
UploadMetricToFactbase(
  data,
  token,
  mode = "replace_all",
  aggregation = "sum",
  definition = NULL,
  hyperlink = NULL,
  update_key = NULL,
  save_failed_json_to = NULL
)
}
\arguments{
\item{data}{A data.frame with at least three columns, being (in order) for
* measurements (must be numeric).  The column name will name the metric.
* date/time (Data Science to specify the formats that normal users would expect to be supported
  in where there will be a lot of data; supply automatic conversion if you think that is
  reasonable).  This column will be identified using its name, which must be `_When`.
* dimension 1 (coerced to character).  The column name will be used for the dimension name.
* …
* dimension n}

\item{token}{A guid that identifies and authenticates the request.  Talk to Oliver if you need
one of these.}

\item{mode}{One of "replace_all", "append" or "append_or_update" See comments for
FactPostUpdateType.}

\item{aggregation}{One of "none", "minimum", "maximum", "sum", "average", "first", "last".}

\item{definition}{A detailed explanation of the meaning and derivation of the metric.}

\item{hyperlink}{A link to a web page where more can be read about the metric.}

\item{update_key}{The name of a column that can be used to update the data, when `mode` is
"append_or_update".  Data in this column must be unique, which implies some sort of aggregation
for date/time data.}

\item{save_failed_json_to}{If set then the JSON for this request will be saved to the named file
in your Displayr Drive.  This is helpful when trying to reproduce a problem for debugging.}

\item{test_expected_json}{Character vector of length 1. If non-NULL then this method will _not_
contact the Factbase server, and will instead just verify that the JSON it would have sent 
matches the JSON provided.  This is to support unit tests.}
}
\value{
The value of `data` that was passed in, so caller can see data uploaded if this is the
  last call in R code.
}
\description{
Upload a metric to Factbase.
}
