% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factbase.R
\name{UploadRelationshipToFactbase}
\alias{UploadRelationshipToFactbase}
\title{Upload a relationship to Factbase.}
\usage{
UploadRelationshipToFactbase(data, token, mode = "replace_all")
}
\arguments{
\item{data}{A data.frame with at least two columns, each of which should be coerced to character
vectors.  The first column is the dimension we are mapping from.  Subsequent columns contain
labels in dimensions that we are mapping to.  The names of these columns to be used as the
names of these dimensions.}

\item{token}{A guid that identifies and authenticates the request.  Talk to Oliver if you need
one of these.}
}
\value{
The value of `data` that was passed in, so caller can see data uploaded if this is the
  last call in R code
}
\description{
Upload a relationship to Factbase.
}
