% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{uploadStandardR}
\alias{uploadStandardR}
\title{Upload a Standard R file from the Q-Wiki-Scripts repo}
\usage{
uploadStandardR(standard.r.input = ".", filename = NULL, upload = TRUE)
}
\arguments{
\item{standard.r.input}{A string containing the file path to use. This needs to be either
\itemize{
\item A path to a directory containing two files, a single R and JS source file respectively.
  E.g. the Calculation features.
\item A path containing the basename of the Standard R page. Suitable for standard R pages
  where a directory contains multiple R and JS files.
}}

\item{filename}{A string containing the desired name of the output script.
Defaults to \code{NULL} and will deduce the name based off the directory structure.}

\item{upload}{A logical whether to upload the script or not (useful to check the script is good before upload).}
}
\description{
Take either a directory or Standard R page basename from the
  Q-Wiki-Scripts repo and upload it to the users custom rscript via Displayr Drive.
}
