% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel.R
\name{DownloadXLSX}
\alias{DownloadXLSX}
\title{Download and read Excel (.xlsx) file}
\usage{
DownloadXLSX(url, sheet = 1, want.data.frame = FALSE, want.factors = TRUE,
  want.col.names = TRUE, want.row.names = FALSE, us.format = TRUE)
}
\arguments{
\item{url}{Link to XLSX file.}

\item{sheet}{Sheet of excel workbook to read from. By default only the first worksheet is read}

\item{want.data.frame}{Whether to return a data frame instead of a matrix or vector. Note that unlike \link[flipTransformations]{ParseEnteredData}, the default is \code{TRUE}, because imported data files are expected to be larger than cut-and-paste data.}

\item{want.factors}{Whether a text variable should be converted to a factor in a data frame.}

\item{want.col.names}{Whether to interpret the first row as column names in a data frame.}

\item{want.row.names}{Whether to interpret the first col as row names in a data frame.}

\item{us.format}{Whether to use the US convention when parsing dates in a data frame.}
}
\description{
Download and read Excel (.xlsx) file
}
