% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel.R
\name{DownloadXLSX}
\alias{DownloadXLSX}
\title{Download and read Excel (.xlsx) file}
\usage{
DownloadXLSX(url, sheet = 1)
}
\arguments{
\item{url}{Link to XLSX file.}

\item{sheet}{Sheet of excel workbook to read from. By default only the first worksheet is read}
}
\description{
Download and read Excel (.xlsx) file
}
