% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactbaseWidget.R
\name{FactBaseMetricWidget}
\alias{FactBaseMetricWidget}
\title{The R code for the standard R item *Upload to Factbase*, which provides a
zero code way to upload a metric using the Factbase R API.}
\usage{
FactBaseMetricWidget(
  factbase.token = "",
  mode = "Replace all",
  aggregation = "None",
  time_aggregation = NULL,
  period_type = NULL,
  definition = NULL,
  hyperlink = NULL,
  owner = NULL,
  do.upload = TRUE,
  selection.type = "Table",
  input.table = NULL,
  make.dummy.metric = NULL,
  metric.variables = NULL,
  metric.column.names = "",
  date.column.name = "",
  date.variable = NULL,
  dimension.column.names = "",
  dimension.variables = NULL,
  output.type = "Data Summaries",
  start.date = "",
  time.zone = "Australia/Sydney",
  update.period = "days",
  update.frequency = 1,
  us.format = FALSE
)
}
\arguments{
\item{factbase.token}{See above}

\item{mode}{See above}

\item{aggregation}{See above}

\item{time_aggregation}{See above}

\item{period_type}{See above}

\item{definition}{See above}

\item{hyperlink}{See above}

\item{owner}{See above}

\item{do.upload}{See above}

\item{selection.type}{See above}

\item{input.table}{See above}

\item{make.dummy.metric}{See above}

\item{metric.variables}{See above}

\item{metric.column.names}{See above}

\item{date.column.name}{See above}

\item{date.variable}{See above}

\item{dimension.column.names}{See above}

\item{dimension.variables}{See above}

\item{output.type}{See above}

\item{start.date}{See above}

\item{time.zone}{See above}

\item{update.period}{See above}

\item{update.frequency}{See above}

\item{us.format}{See above}
}
\value{
An HTMLwidget that shows summary details about the upload.
}
\description{
Parameters come out of the R UI JavaScript code in
https://github.com/Displayr/factbase/blob/master/displayr/Upload%20to%20Factbase.RScript
See that code to understand the meaning of those controls.
}
