% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeocodeIPs.R
\name{GeocodeIPs}
\alias{GeocodeIPs}
\title{GeocodeIP addresses}
\usage{
GeocodeIPs(ips)
}
\arguments{
\item{ips}{Character  \code{\link{vector}} of IP addresses.}
}
\value{
A \code{\link{data.frame}} with columns
\itemize{
\item \code{ips} - the original input ips
\item \code{status} - the result of the geocoding query
\item \code{country_name} - the name of the country where the IP is located
\item \code{city} - the name of the city where the IP is located
}
}
\description{
Returns the country and continent of a vector of IPv4 or IPv6 strings.
}
\examples{
GeocodeIPs(c("123.51.111.134", "216.27.61.137", "2001:780:53d2::1"))
}
