% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeocodeIPs.R
\name{GeocodeIPs}
\alias{GeocodeIPs}
\title{GeocodeIP addresses}
\usage{
GeocodeIPs(ips)
}
\arguments{
\item{ips}{Character  \code{\link{vector}} of IP addresses.}
}
\value{
A \code{\link{data.frame}} with columns
\itemize{
\item \code{ips} - the original input ips
\item \code{continent_name} - the name of the continent where the IP is located
\item \code{country_name} - the name of the country where the IP is located
\item \code{country_code} - the 2 letter country code (ISO Alpha-2)
}
}
\description{
Returns the country and continent of a vector of IPv4 or IPv6 strings.
}
\details{
Uses the \href{https://www.maxmind.com/en/home}{MaxMind} database from
   the \code{\link[rgeolocate:maxmind]{rgeolocate}} package.
   Returns \code{\link{NA}} when no data is available.
}
\examples{
GeocodeIPs(c("123.51.111.134", "216.27.61.137", "2001:780:53d2::1"))
}
