% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeocodeIPs.R
\name{GeocodeIPs}
\alias{GeocodeIPs}
\title{GeocodeIP addresses}
\usage{
GeocodeIPs(ips)
}
\arguments{
\item{ips}{Character  \code{\link{vector}} of IP addresses.}
}
\value{
A \code{\link{data.frame}} with columns
\itemize{
\item \code{ips} - the original input ips
\item \code{continent_name} - the name of the continent where the IP is located
\item \code{country_name} - the name of the country where the IP is located
\item \code{country_code} - the 2 letter country code (ISO Alpha-2)
}
}
\description{
Returns the country and continent of a vector of IPv4 or IPv6 strings.
}
\details{
Uses the \href{https://lite.ip2location.com/ip2location-lite}{IP2Location LITE}
   or \href{https://www.ip2location.com/database/ip2location}{IP2Location} databases
   from IP2Location to resolve country locations from IP addresses.
   The database must be downloaded and the database path set as the environment variable
   \code{IP2LOCATION_DB_PATH} before use. The database can be downloaded from
   \href{https://lite.ip2location.com/ip2location-lite}{IP2Location LITE} or
   \href{https://www.ip2location.com/database/ip2location}{IP2Location}.
   To interact with the database a python environment is required that has the
   \href{https://pypi.org/project/IP2Location/}{IP2Location} python package installed.
   If an invalid IP address or the information is not available in the database,
   the \code{country_name} and \code{country_code} columns will be set to \code{NA}.
   The LITE database should achieve an accuracy at the country level of 98\%. If
   you need more accuracy, you can use the commercial license with the IP2Location
   database. More details on the differences between the databases can be found
   at \href{https://lite.ip2location.com/edition-comparison}{IP2Location comparisons}.
}
\examples{
\dontrun{
GeocodeIPs(c("123.51.111.134", "216.27.61.137", "2001:780:53d2::1"))
}
}
