% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{getClientId}
\alias{getClientId}
\title{Gets the client Id (project id) from the environment. Throws an error if not found.}
\usage{
getClientId()
}
\value{
The client id as a string
}
\description{
Gets the client Id (project id) from the environment. Throws an error if not found.
}
