% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{GetCompanySecret}
\alias{GetCompanySecret}
\title{Gets company secret from the environment. Throws an error if not found.}
\usage{
GetCompanySecret()
}
\value{
Company secret token as a string.
}
\description{
Gets company secret from the environment. Throws an error if not found.
}
