% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{QDeleteFiles}
\alias{QDeleteFiles}
\title{Deletes a set of objects}
\usage{
QDeleteFiles(filenames, company.token = getCompanySecret())
}
\arguments{
\item{filenames}{collection of character strings. Names of the files to delete.}

\item{company.token}{Use this if you need to read from a different company's Displayr Cloud Drive.  You need to contact Support to get this token.}
}
\value{
NULL invisibly. Called for the purpose of deleting data
and assumed to succeed if no errors are thrown.
}
\description{
Deletes a list of objects by filename from the Displayr cloud drive
}
