% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{QFileExists}
\alias{QFileExists}
\title{Check if a file exists}
\usage{
QFileExists(filename, show.warning = TRUE)
}
\arguments{
\item{filename}{character string. Name of the file to search for.}

\item{show.warning}{logical scalar. Whether to show a warning when the file
does not exist.}
}
\value{
TRUE if the file exists, otherwise FALSE.
}
\description{
Check whether a file of a given name exists in the Displayr Cloud Drive.
}
