% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{QFileOpen}
\alias{QFileOpen}
\title{Opens a Connection}
\usage{
QFileOpen(filename, open = "r", blocking = TRUE,
  encoding = getOption("encoding"), raw = FALSE,
  method = getOption("url.method", "default"))
}
\arguments{
\item{filename}{character string. Name of file to be opened}

\item{open}{character string. A description of how to open the connection}

\item{blocking}{logical. Whether or not the connection blocks can be specified}

\item{encoding}{character string. The name of the encoding to be assumed.}

\item{raw}{logical. Whether this connection should be treated as a byte stream}

\item{method}{character string. See documentation for connections}
}
\value{
A curl connection (read) or a file connection (write)
}
\description{
Opens a connection for either reading OR writing to a file of a given name.
In the reading case, it opens a stream to the file in the Data Mart. 
In the writing case, it opens a temporary file for writing to and uploads this to the Data Mart on close.
Note that writing to a file which already exists will overwrite that file's contents.
}
