% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{QLoadData}
\alias{QLoadData}
\title{Loads an object}
\usage{
QLoadData(filename, company.token = NA)
}
\arguments{
\item{filename}{character string. Name of the file to be opened from the Data Mart.}

\item{company.token}{Use this if you need to read from a different Displayr company's data mart.  You need to contact Support to get this token.}
}
\value{
An R object
}
\description{
Loads an *.rds file from the data mart and converts this to an R object.
}
