% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{QLoadData}
\alias{QLoadData}
\title{Loads an object}
\usage{
QLoadData(filename, company.token = NA, ...)
}
\arguments{
\item{filename}{character string. Name of the file to be opened from the Displayr Cloud Drive.}

\item{company.token}{Use this if you need to read from a different company's Displayr Cloud Drive.  You need to contact Support to get this token.}

\item{...}{Other parameters to pass to read.csv.}
}
\value{
An R object
}
\description{
Loads an *.csv, *.rds or *.sav file from the Displayr Cloud Drive and
converts this to an R object.
}
