% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{QSaveData}
\alias{QSaveData}
\title{Save an object}
\usage{
QSaveData(object, filename, compression.file.size.threshold = NULL, ...)
}
\arguments{
\item{object}{object. The object to be uploaded.}

\item{filename}{character string. Name of the file to be written to.}

\item{compression.file.size.threshold}{numeric scalar. Files of size
(in bytes) larger than this value will be compressed into a zip file.
Defaults to NULL, in which case no compression occurs.}

\item{...}{Other parameters to pass to \code{\link{write.csv}}, \code{\link{saveRDS}},
\code{\link{write.xlsx}}, or \code{\link{write_sav}}.}
}
\value{
NULL invisibly. Called for the purpose of uploading data
and assumed to succeed if no errors are thrown.
}
\description{
Saves an object to the Displayr Cloud Drive without any transformation.
Filename string must have a .csv, .rds, .xlsx, pptx, gif, or .sav extension.
}
\note{
Saving to Powerpoint .pptx files is only possible for rpptx objects created using
the \code{officer} package.

Saving to .gif file is only supported for \code{"gganim"} objects created using
\code{gganimate}.

When saving to .xlsx file, \code{object} is first coerced to a data.frame using
\code{\link{as.data.frame}}.
}
