% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{QSaveData}
\alias{QSaveData}
\title{Save an object}
\usage{
QSaveData(object, filename)
}
\arguments{
\item{object}{object. The object to be uploaded.}

\item{filename}{character string. Name of the file to be written to.}
}
\value{
NULL invisibly. Called for the purpose of uploading data
and assumed to succeed if no errors are thrown.
}
\description{
Saves an object to the Displayr Cloud Drive without any transformation.
Filename string must have a .csv or .rds extension.
}
