% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMart.R
\name{StopNotDisplayr}
\alias{StopNotDisplayr}
\title{Error when someone tries to use this package outside of Displayr.}
\usage{
StopNotDisplayr()
}
\value{
Throws an error.
}
\description{
Error when someone tries to use this package outside of Displayr.
}
