% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factbase.R
\name{UploadMetricToFactbase}
\alias{UploadMetricToFactbase}
\title{Upload a metric to Factbase.}
\usage{
UploadMetricToFactbase(
  data,
  token,
  mode = "replace_all",
  aggregation = "sum",
  definition = NULL,
  hyperlink = NULL
)
}
\arguments{
\item{data}{A data.frame with at least three columns, being (in order) for 
* measurements (must be numeric).  The column name will name the metric. 
* date/time (Data Science to specify the formats that normal users would expect to be supported
  in where there will be a lot of data; supply automatic conversion if you think that is
  reasonable).  The column name is unimportant.
* dimension 1 (coerced to character).  The column name will be used for the dimension name.
* …
* dimension n}

\item{token}{A guid that identifies and authenticates the request.  Talk to Oliver if you need
one of these.}

\item{mode}{One of "replace_all", "append" or "append_or_update" See comments for
FactPostUpdateType.}

\item{aggregation}{One of "none", "minimum", "maximum", "sum", "average", "first", "last".}

\item{definition}{A detailed explanation of the meaning and derivation of the metric.}

\item{hyperlink}{A link to a web page where more can be read about the metric.}
}
\value{
The value of `data` that was passed in, so caller can see data uploaded if this is the
  last call in R code
}
\description{
Upload a metric to Factbase.
}
