% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factbase.R
\name{UploadRelationshipToFactbase}
\alias{UploadRelationshipToFactbase}
\title{Upload a relationship to Factbase.}
\usage{
UploadRelationshipToFactbase(
  data,
  token,
  mode = "replace_all",
  definition = NULL,
  hyperlink = NULL,
  owner = NULL,
  save_failed_json_to = NULL,
  test_return_json = FALSE
)
}
\arguments{
\item{data}{A data.frame with at least two columns, each of which should be coerced to character
vectors.  The first column is the dimension we are mapping from.  Subsequent columns contain
labels in dimensions that we are mapping to.  The names of these columns to be used as the
names of these dimensions.}

\item{token}{A guid that identifies and authenticates the request.  Talk to Oliver if you need
one of these.}

\item{mode}{(optional) One of "replace_all", "append" or "append_or_update" See comments for
FactPostUpdateType.}

\item{definition}{(optional) A detailed explanation of the meaning and derivation of the data.}

\item{hyperlink}{(optional) A link to a web page where more can be read about the data.
Preferably this is a link into the system that calls this function.}

\item{owner}{The name (usually an email address) of whoever should be contacted to deal with problems
or questions about this data.}

\item{save_failed_json_to}{(optional) If set then the JSON for this request will be saved to the named file
in your Displayr Drive.  This is helpful when trying to reproduce a problem for debugging.}

\item{test_return_json}{(optional) For testing only.  Ignore.}
}
\value{
The value of `data` that was passed in, so caller can see data uploaded if this is the
  last call in R code.
}
\description{
Upload a relationship to Factbase.
}
