% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factbase.R
\name{UploadRelationshipToFactbase}
\alias{UploadRelationshipToFactbase}
\title{Upload a relationship to Factbase.}
\usage{
UploadRelationshipToFactbase(
  data,
  token,
  mode = "replace_all",
  save_failed_json_to = NULL
)
}
\arguments{
\item{data}{A data.frame with at least two columns, each of which should be coerced to character
vectors.  The first column is the dimension we are mapping from.  Subsequent columns contain
labels in dimensions that we are mapping to.  The names of these columns to be used as the
names of these dimensions.}

\item{token}{A guid that identifies and authenticates the request.  Talk to Oliver if you need
one of these.}

\item{mode}{One of "replace_all", "append" or "append_or_update" See comments for
FactPostUpdateType.}

\item{save_failed_json_to}{If set then the JSON for this request will be saved to the named file
in your Displayr Drive.  This is helpful when trying to reproduce a problem for debugging.}

\item{test_expected_json}{Character vector of length 1. If non-NULL then this method will _not_
contact the Factbase server, and will instead just verify that the JSON it would have sent 
matches the JSON provided.  This is to support unit tests.}
}
\value{
The value of `data` that was passed in, so caller can see data uploaded if this is the
  last call in R code.
}
\description{
Upload a relationship to Factbase.
}
