% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factbase.R
\name{UploadTableToFactbase}
\alias{UploadTableToFactbase}
\title{Upload a table of raw data to Factbase.}
\usage{
UploadTableToFactbase(
  table_name,
  data,
  token,
  mode = "replace_all",
  na_columns = NULL,
  test_return_json = FALSE
)
}
\arguments{
\item{table_name}{The name to use to refer to this data in Factbase.}

\item{data}{A data.frame containing columns of data.  Character, factor (converted to character),
numeric, boolean (converted to character) and date/time (`Date` or `POSIXt`) columns are acceptable.}

\item{token}{A guid that identifies and authenticates the request.  Talk to Oliver if you need
one of these.}

\item{mode}{(optional) One of "replace_all", "append" or "append_or_update" See comments for
FactPostUpdateType.}

\item{na_columns}{(Optional) If set then this should be a character vector naming the
columns that may contain NAs, which will be converted into nulls int the resultant table.}

\item{test_return_json}{(Optional) For testing only.  Ignore.}
}
\value{
The value of `data` that was passed in, so caller can see data uploaded if this is the
  last call in R code.
}
\description{
Upload a table of raw data to Factbase.
}
